Save user configuration settings for gpttools

This function saves the user's configuration settings for gpttools service to a YAML file.

Usage

save_user_config(
  service = "openai",
  model = "gpt-4-turbo-preview",
  task = "Permissive Chat",
  local_embed = FALSE,
  openai_embed_model = "text-embedding-3-small",
  local_embed_model = "BAAI/bge-small-en-v1.5",
  k_context = 4,
  k_history = 4,
  save_history = TRUE,
  add_context = "sometimes",
  sources = "All",
  run_code = FALSE,
  persist = TRUE
)

Arguments

service

The name of the service to use, default is "openai".

model

The model to use, default is "gpt-4-1106-preview".

task

The task to perform, default is "Permissive Chat".

local_embed

Whether to use local embedding model. Default is FALSE.

openai_embed_model

The OpenAI embeddings model to use, default is "text-embedding-3-small".

local_embed_model

The local embeddings model to use, default is "BAAI/bge-small-en-v1.5".

k_context

The amount of context to keep, default is 4.

k_history

The amount of history to keep, default is 4.

save_history

Logical indicating whether history should be saved, default is TRUE.

add_context

Whether to add context to the query. Options are "always", "sometimes", and "never". The default is "sometimes".

sources

The sources to use, default is "All".

run_code

Whether to execute generated code with reprex::reprex(), default is FALSE.

persist

Logical indicating whether to persist the settings, default is TRUE.

Value

Invisible NULL.